Update float to double

+++ code/include/group.hpp
@@ -24,7 +24,7 @@ public:
 
     }
 
-    bool intersect(const Ray &r, Hit &h, float tmin) override {
+    bool intersect(const Ray &r, Hit &h, double tmin) override {
         bool ret = false;
         for (auto object : group)
             ret |= object->intersect(r, h, tmin);
